refactor: Move Game component to routes/hello/components directory
The Game component and all of its related elements have been moved from interface directory to routes/hello/components directory. This includes the component itself and its associated modules involving files like key, polygon, snake-placeholder etc. Additionally, the import paths have been updated accordingly.
