+++ src/routes/hello/components/snake-placeholder.tsx
